Run all Rails tests in 1 GHA step
This saves us from bundle installing all of the appraisal versions in
separate CI steps.  Ultimately it makes CI run faster, but we will have
to dig into CI logs to see which version of Rails might have produced a
failure.  The Rails test suite is very tiny, though, so this shouldn't
be too painful.